+++ core/src/ten_rust/tests/test_case/graph/mod.rs
@@ -6,6 +6,7 @@
 //
 mod exposed_message;
 mod flatten_integration;
+mod graph_info;
 mod import_uri;
 mod subgraph;
